week-06/day1/notes/noun_project/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
idna==3.4
oauthlib==3.2.2
python-dotenv==1.0.0
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
urllib3==1.26.15
